adding ncr

+++ ModInverse.cpp
@@ -1,17 +1,17 @@
-int gcdExtended(int a, int b, int *x, int *y){
+ll gcdExtended(ll a, ll b, ll *x, ll *y){
     if (a == 0){ 
         *x = 0, *y = 1; 
         return b; 
     } 
-    int x1, y1;
-    int gcd = gcdExtended(b%a, a, &x1, &y1); 
+    ll x1, y1;
+    ll gcd = gcdExtended(b%a, a, &x1, &y1); 
     *x = y1 - (b/a) * x1; 
     *y = x1;
     return gcd; 
 }
-int modInverse(int a, int m){ 
-    int x, y; 
-    int g = gcdExtended(a, m, &x, &y); 
-    int res = (x%m + m) % m; 
+ll modInverse(ll a, ll m){ 
+    ll x, y; 
+    ll g = gcdExtended(a, m, &x, &y); 
+    ll res = (x%m + m) % m; 
     return res;
 }
